% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{compute_mean}
\alias{compute_mean}
\title{compute_mean}
\usage{
compute_mean(data.matrix)
}
\arguments{
\item{data.matrix}{A matrix with gene expression values, rows denote genes and columns denote replicate samples. Alternatively a vector of gene expression values.}
}
\value{
This function returns a vector with the mean expression values.
}
\description{
This function calculates average expression values of multiple replicate samples
}
\details{
This function calculates average expression values of multiple replicate samples using rowMeans. If the data has no replicates (a single vector or a matrix with only one column) the original values are returned.
}
\examples{
dd<-matrix(sample(1:10, 30, replace=TRUE), 10, 3)
rownames(dd)<-c("CRYAA", "CRYAB", "CRYBB3", "PAX6", "SOX2", "PROX1", "SIX3", "CRIM1", "CRYBB2", "BMP7")
compute_mean(dd)

}
