% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{display_pathway_ranking_metric}
\alias{display_pathway_ranking_metric}
\title{display_pathway_ranking_metric}
\usage{
display_pathway_ranking_metric(pathway.ranking.metric)
}
\arguments{
\item{pathway.ranking.metric}{The ranking metric result returned by 'get_pathway_ranking_metric' function.}
}
\description{
This function plots the pathway ranking metric for each cell type or tissue in a 2D plane.
}
\details{
This function plots the pathway ranking metric for each cell type or tissue in a 2D plane.
}
\examples{
## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
## Get active pathway ranking metric (i.e., activity score and number of downstream transcription factors)
ROR1.active.pathway.ranking.metric<-get_pathway_ranking_metric(active.pathway.path = ROR1.active.pathway, processed.query.data = ROR1.processed.data, high.exp.th = 7)
## Plot the ranking metric result (i.e., activity score and number of downstream transcription factors) in a 2D plane
display_pathway_ranking_metric(pathway.ranking.metric = ROR1.active.pathway.ranking.metric)
## To separate the top ranked pathways we can do this
abline(v=45, h=0.2, lty=2, col="black")

}
