% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{display_top_ranked_pathways}
\alias{display_top_ranked_pathways}
\title{display_top_ranked_pathways}
\usage{
display_top_ranked_pathways(pathway.ranking.metric, top.n.pathway = 50)
}
\arguments{
\item{pathway.ranking.metric}{The ranking metric result returned by 'get_active_pathway_ranking_metric' function.}

\item{top.n.pathway}{An integer number of how many top ranked pathways will be displayed. Default is 50.}
}
\description{
This function displays the sorted top n ranked pathways for each cell type or tissue in a barplot.
}
\details{
This function displays the sorted top n ranked pathways for each cell type or tissue in a barplot.
}
\examples{
## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
## Get active pathway ranking metric
ROR1.active.pathway.ranking.metric<-get_active_pathway_ranking_metric(active.pathway.path = ROR1.active.pathway, processed.query.data = ROR1.processed.data, high.exp.th = 7)
## Display top n pathways
display_top_ranked_pathways(pathway.ranking.metric = ROR1.active.pathway.ranking.metric)

}
