% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{generate_pathway_path}
\alias{generate_pathway_path}
\title{generate_pathway_path}
\usage{
generate_pathway_path(ppi.result, housekeeping.gene, max.path.length = 7)
}
\arguments{
\item{ppi.result}{The list object returned by the function combine_mm_hs_ppi.}

\item{housekeeping.gene}{A vector consisting of housekeeping genes. This data is loaded automatically with the package. This data was generated using the gene expression profiles of different cell types and/or tissues from the ENCODE human and mouse project.}

\item{max.path.length}{An integer number indicating the maximum path length. Default is 7.}
}
\value{
This function returns a list consisting of the pathway path data that will be used as background data for the SPAGI package.
}
\description{
This function generates the background pathway path data using the list object returned by the function combine_mm_hs_ppi.
}
\details{
This function generates the background pathway path data using the list object returned by the function combine_mm_hs_ppi.
}
\examples{
## Need two folder at working directory for downloading stringdb files for each species - stringdb_mouse, stringdb_human.
## It takes some time to download the data, and then can reuse the downloaded data.
## Here we will use RP.protein, KN.protein, TF.protein protein parameters. These data are automatically loaded with the package. You can modify these parameters.
## We will generate PPI data for two species - "mmusculus" and "hsapiens" by calling the function get_ppi_for_molecules two times.
## Then we will combine these two PPI data sets by using the combine_mm_hs_ppi function that will be used to generate the pathway path data.
## Finally we will generate the pathway path data using the combined PPI data

## Get PPI data for the protein molecules of species "mmusculus".
mm.ppi<-get_ppi_for_molecules(RP.protein, KN.protein, TF.protein, species="mmusculus")
## Get PPI data for the protein molecules of species "hsapiens".
hs.ppi<-get_ppi_for_molecules(RP.protein, KN.protein, TF.protein, species="hsapiens")
## Now combine and get the filtered PPI and the RP and TF proteins of the combined filtered PPI
comb.ppi.result<-combine_mm_hs_ppi(mm.ppi, hs.ppi, RP.protein, KN.protein, TF.protein)
##Generate the pathway path data using the comb.ppi.result and housekeeping.gene data sets
pathway.path<-generate_pathway_path(ppi.result=comb.ppi.result, housekeeping.gene)
head(summary(pathway.path))

}
