% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{get_pathway_downstream_tf_number}
\alias{get_pathway_downstream_tf_number}
\title{get_pathway_downstream_tf_number}
\usage{
get_pathway_downstream_tf_number(active.pathway.path)
}
\arguments{
\item{active.pathway.path}{A list of active pathway path data for each cell/tissue type as returned by the function 'identify_active_pathway_path'.}
}
\value{
This function returns a list of number of downstrean transcription factors for each cell/tissue type.
}
\description{
This function calculates the number of downstrean transcription factors of each active pathway for each cell/tissue type. It uses the active pathway path object to calculate the number of downstrean transcription factor.
}
\details{
This function calculates the number of downstrean transcription factors of each active pathway for each cell/tissue type. It uses the active pathway path object to calculate the number of downstrean transcription factor.
}
\examples{
## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format. Also we have already made the replicate names same for the data.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
## Get the number of downstream transcription factors of the active pathways
ROR1.pathway.downstream.tf.count<-get_pathway_downstream_tf_number(active.pathway.path = ROR1.active.pathway)
head(ROR1.pathway.downstream.tf.count$ROR1_LEC)

}
