% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{get_pathway_ranking_metric}
\alias{get_pathway_ranking_metric}
\title{get_pathway_ranking_metric}
\usage{
get_pathway_ranking_metric(active.pathway.path, processed.query.data,
  high.exp.th)
}
\arguments{
\item{active.pathway.path}{A list of active pathway path data for each cell/tissue type as returned by the function 'identify_active_pathway_path'.}

\item{processed.query.data}{A list with expressed query data where each sublist corresponds for each cell/tissue type as returned by the function 'preprocess_querydata'.}

\item{high.exp.th}{A high expression threshold value for the processed query data. It is used to get active (i.e., highly expressed) molecule proportion for each pathway path.}
}
\value{
This function returns a list of sublist with pathway activity score and the number of downstream transcription factors for each cell/tissue type.
}
\description{
This function generates pathway activity score and number of downstream transcription factors of the active pathways for each cell/tissue type. It uses active pathway path and processed query data with a high expression threshold to generate the activity score. However, it uses only the active pathway path data to calculate the number of downstream transcription factors.
}
\details{
This function generates pathway activity score and number of downstream transcription factors of the active pathways for each cell/tissue type. It uses active pathway path and processed query data with a high expression threshold to generate the activity score. However, it uses only the active pathway path data to calculate the number of downstream transcription factors.
}
\examples{
## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format. Also we have already made the replicate names same for the data.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
## Get active pathway ranking metric (i.e., activity score and number of downstream transcription factors)
ROR1.active.pathway.ranking.metric<-get_pathway_ranking_metric(active.pathway.path = ROR1.active.pathway, processed.query.data = ROR1.processed.data, high.exp.th = 7)
head(ROR1.active.pathway.ranking.metric$activity.score$ROR1_LEC)
head(ROR1.active.pathway.ranking.metric$downstream.tf.count$ROR1_LEC)

}
