% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{get_ppi_for_molecules}
\alias{get_ppi_for_molecules}
\title{get_ppi_for_molecules}
\usage{
get_ppi_for_molecules(RP.protein, KN.protein, TF.protein, species,
  score = 700)
}
\arguments{
\item{RP.protein}{A vector containg the receptor (RP) proteins.}

\item{KN.protein}{A vector containg the kinase (KN) proteins.}

\item{TF.protein}{A vector containg the transcription factor (TF) proteins.}

\item{species}{The species name, either "hsapiens" or "mmusculus".}

\item{score}{An interger value for the STRINGdb PPI score threshold cutoff. Default is 700.}
}
\value{
This function returns a data frame of the PPI for the molecules.
}
\description{
This function gets the PPI data from STRINGdb for the protein molecules provided.
}
\details{
This function gets the PPI data from STRINGdb for the protein molecules provided.
}
\examples{
## Need two folder at working directory for downloading stringdb files for each species - stringdb_mouse, stringdb_human.
## It takes some time to download the data, and then can reuse the downloaded data.
## Here we will use RP.protein, KN.protein, TF.protein protein parameters. These data are automatically loaded with the package. You can modify these parameters.
## And we will use the species as "mmusculus".

## Get PPI data for the protein molecules of species "mmusculus".
mm.ppi<-get_ppi_for_molecules(RP.protein, KN.protein, TF.protein, species="mmusculus")
head(mm.ppi)

}
