% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housekeeping.gene}
\alias{housekeeping.gene}
\title{Housekeeping gene data}
\format{An object of vector containing housekeeping genes that is used as background data.}
\usage{
data(housekeeping.gene)
}
\description{
Housekeeping gene used as background data, generated using gene expression profiles of different cells/tissues from ENCODE project.
}
\examples{
head(housekeeping.gene)
}
\keyword{datasets}
