% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\name{identify_active_pathway_path}
\alias{identify_active_pathway_path}
\title{identify_active_pathway_path}
\usage{
identify_active_pathway_path(pathway.path, processed.query.data)
}
\arguments{
\item{pathway.path}{A list with pathway path data where each sublist denotes a path of the pathway. This is used as background data.}

\item{processed.query.data}{A list with expressed query data where each sublist corresponds for each cell/tissue type.}
}
\value{
This function returns a list of pathways where each sublist denote a path of the pathway.
}
\description{
This function identifies active pathway paths for RNA-seq gene expression profile. It utilizes background pathway path data to identify the active pathway paths.
}
\details{
This function identifies active pathway paths for RNA-seq gene expression profile. It utilizes background pathway path data to identify the active pathway paths.
}
\examples{
## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
head(ROR1.active.pathway$ROR1_LEC$FGFR1)

}
