% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KN.protein}
\alias{KN.protein}
\title{Kinase protein data}
\format{An object of vector containing kinase proteins.}
\usage{
data(KN.protein)
}
\description{
Kinase protein, used as middle components of pathway paths.
}
\examples{
head(KN.protein)
}
\keyword{datasets}
