% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.path}
\alias{pathway.path}
\title{Pathway path data}
\format{An object of list containing pathway paths that is used as background data.}
\usage{
data(pathway.path)
}
\description{
Pathway path used as background data, generated using STRINGdb PPI.
}
\examples{
head(summary(pathway.path)
}
\keyword{datasets}
