% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROR1.data}
\alias{ROR1.data}
\title{ROR1 data}
\format{An object of matrix containing Count Per Million (CPM) and log2 normalized gene expression value of LEC.}
\usage{
data(ROR1.data)
}
\description{
Data from the ROR1+ cells differentiated from human pluripotent stem cells,
contains transcriptomic profiles data of lens epithelial cell (LEC).
The cell has 2 biological replicates.
}
\examples{
head(ROR1.data)
}
\references{
Murphy P et al, Development 2018 Jan 9; 145(1)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29217756}{PubMed})
}
\keyword{datasets}
