% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RP.protein}
\alias{RP.protein}
\title{Receptor protein data}
\format{An object of vector containing receptor proteins.}
\usage{
data(RP.protein)
}
\description{
Receptor protein, used as source components of pathway paths.
}
\examples{
head(RP.protein)
}
\keyword{datasets}
