% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF.protein}
\alias{TF.protein}
\title{Transcription factor protein data}
\format{An object of vector containing Transcription factor proteins.}
\usage{
data(TF.protein)
}
\description{
Transcription factor protein, used as target components of pathway paths.
}
\examples{
head(TF.protein)
}
\keyword{datasets}
