% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi_master.R
\docType{package}
\name{spagi-package}
\alias{spagi-package}
\alias{spagi}
\alias{SPAGI}
\title{SPAGI: Signalling Pathway Analysis for putative Gene regulatory network Identification}
\description{
SPAGI is an R package for active pathway identification for RNA-seq gene expression profiles. This package contains the neccessary R code to run SPAGI as described in "SPAGI: Identification of active signalling pathways using RNA-seq gene expression profiles". SPAGI is implemented within a helpful framework to identify active pathway paths from RNA-seq gene expression profiles.
}
\examples{
## Do a sample analysis using human ocular lens epithelial cell (LEC) RNA-seq gene expression data.

## Here we will use "pathway.path" as background data from the SPAGI repository.
## Also we will use "ROR1.data" as query RNA-seq gene expression data. This data is for ocular lens epithelial cell differentiated from human pluripotent stem cells.
## These data sets are loaded automatically with the package.

## Pre-process the query data (ROR1.data), the data has already been made in CPM and log2 normalized format.
ROR1.processed.data<-preprocess_querydata(cell.tissue.data = ROR1.data, exp.cutoff.th = 1.8)
## Identify active pathway paths of the processed query data
ROR1.active.pathway<-identify_active_pathway_path(pathway.path = pathway.path, processed.query.data = ROR1.processed.data)
## Get active pathway ranking metric
ROR1.active.pathway.ranking.metric<-get_active_pathway_ranking_metric(active.pathway.path = ROR1.active.pathway, processed.query.data = ROR1.processed.data, high.exp.th = 7)
## Display top n pathways
display_top_ranked_pathways(pathway.ranking.metric = ROR1.active.pathway.ranking.metric)

}
\author{
Md Humayun Kabir <humayun.mkabir@gmail.com>
}
